backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
groq
python-multipart
pdfplumber
PyPDF2
python-docx
openpyxl
pandas
Pillow
scikit-learn
numpy
chromadb
sentence-transformers
jinja2
python-dotenv
playwright
